% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zementisTutorials.R
\docType{package}
\name{zementisTutorials-package}
\alias{zementisTutorials}
\alias{zementisTutorials-package}
\title{zementisTutorials: Data Science and Zementis tutorials based on R, Zementis, and PMML}
\description{
The package offers several tutorials in the domain of data science, Zementis, and PMML.
  Students will gain a general understanding of the entire data science process. In addition, they 
  will learn how to operationalize their analytics setup with Zementis Server and PMML.
}
\section{Usage}{

The following courses are currently offered via zementisTutorials and can be executed using the
learnr package:

\strong{De-mystifying Data Science}

The course introduces the data science process and common analytics terminology hands-on using R.
Students will solve a real-world data science problem on their own. The course is completely
self-contained and no prior analytics or R knowledge is required.

\code{learnr::run_tutorial("data_science_introduction", package = "zementisTutorials")}

\strong{Zementis Introduction}

Basic introduction to Zementis Server from an end-user perspective. Students will learn
how to deploy a ML model to Zementis Server and how to score new data using file uploads,
the embedded REST client, an external REST client, and the Zementis Excel Add-In.

\code{learnr::run_tutorial("zementis_introduction", package = "zementisTutorials")}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alex23lemm/zementisTutorials}
  \item Report bugs at \url{https://github.com/alex23lemm/zementisTutorials/issues}
}

}
\author{
\strong{Maintainer}: Alexander Lemm \email{alex23lemm@gmail.com}

}
\keyword{internal}
